django-api/pyproject.toml
[project]
name = "django-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-yasg>=1.21.10",
    "gunicorn>=23.0.0",
    "uvicorn>=0.34.2",
    "uvicorn-worker>=0.3.0",
    "pillow (>=11.2.1,<12.0.0)",
]
